# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/activity_types:
    get:
      tags:
        - ActivityTypes
      summary: List ActivityTypes
      description: Get a paginated list of ActivityTypes with optional filtering and sorting
      operationId: listActivityTypes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityTypeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ActivityTypes
      summary: Create ActivityType
      description: Create a new ActivityType
      operationId: createActivityType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityTypeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/projects:
    get:
      tags:
        - Projects
      summary: List Projects
      description: Get a paginated list of Projects with optional filtering and sorting
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Projects
      summary: Create Project
      description: Create a new Project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    ActivityType:
      type: object
      required:
        - id
        - company_id
        - name
        - billing_rate
        - costing_rate
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        billing_rate:
          type: string
        costing_rate:
          type: string
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreateActivityTypeRequest:
      type: object
      required:
        - company_id
        - name
        - billing_rate
        - costing_rate
        - is_active
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        billing_rate:
          type: string
        costing_rate:
          type: string
        is_active:
          type: boolean
        metadata:
          type: object
...
    ActivityTypeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ActivityType'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Project:
      type: object
      required:
        - id
        - company_id
        - project_name
        - project_type
        - currency
        - status
        - total_costing_amount
        - total_billable_amount
        - total_billed_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        project_name:
          type: string
          maxLength: 200
        project_type:
          type: string
        customer_id:
          nullable: true
          type: string
          format: uuid
        source_so_id:
          nullable: true
          type: string
          format: uuid
        currency:
          type: string
        status:
          type: string
        expected_start_date:
          nullable: true
          type: string
          format: date-time
        expected_end_date:
          nullable: true
          type: string
          format: date-time
        total_costing_amount:
          type: string
        total_billable_amount:
          type: string
        total_billed_amount:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 2000
        metadata:
          type: object
...
    CreateProjectRequest:
      type: object
      required:
        - company_id
        - project_name
        - project_type
        - currency
        - status
        - total_costing_amount
        - total_billable_amount
        - total_billed_amount
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        project_name:
          type: string
          maxLength: 200
        project_type:
          type: string
        customer_id:
          nullable: true
          type: string
          format: uuid
        source_so_id:
          nullable: true
          type: string
          format: uuid
        currency:
          type: string
        status:
          type: string
        expected_start_date:
          nullable: true
          type: string
          format: date-time
        expected_end_date:
          nullable: true
          type: string
          format: date-time
        total_costing_amount:
          type: string
        total_billable_amount:
          type: string
        total_billed_amount:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 2000
        metadata:
          type: object
...
    ProjectList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []